Added BufferedFluentMetric
This acts like FluentMetric in every way except that it doesn't always
immediately send metrics to cloudwatch. Instead, it generally waits
until it has the maximum number of items per batch that cloudwatch
allows (20). Using BufferedFluentMetric should generally result in less
API calls being made to cloudwatch.

The flush() function can be called explicitly at the end of a request to
force all metrics to be sent immediately. This is necessary for Lambda
requests, because the VM may terminate with unsent metrics still
buffered.